fix: resolve broken references
