+++ packages/react-query/src/__tests__/QueryResetErrorBoundary.test.tsx
@@ -1,10 +1,9 @@
-import { waitFor, fireEvent } from '@testing-library/react'
+import { fireEvent, waitFor } from '@testing-library/react'
 import { ErrorBoundary } from 'react-error-boundary'
 import * as React from 'react'
 
-import { sleep, queryKey, createQueryClient } from './utils'
-import { renderWithClient } from './utils.test'
-import { useQuery, QueryCache, QueryErrorResetBoundary } from '..'
+import { createQueryClient, queryKey, renderWithClient, sleep } from './utils'
+import { QueryCache, QueryErrorResetBoundary, useQuery } from '..'
 
 // TODO: This should be removed with the types for react-error-boundary get updated.
 declare module 'react-error-boundary' {
