+++ packages/react-query/src/__tests__/useIsMutating.test.tsx
@@ -1,10 +1,13 @@
-import { waitFor, fireEvent } from '@testing-library/react'
+import { fireEvent, waitFor } from '@testing-library/react'
 import * as React from 'react'
 import { useIsMutating } from '../useIsMutating'
 import { useMutation } from '../useMutation'
-import { createQueryClient, setActTimeout, sleep } from './utils'
-
-import { renderWithClient } from './utils.test'
+import {
+  createQueryClient,
+  renderWithClient,
+  setActTimeout,
+  sleep,
+} from './utils'
 import { ErrorBoundary } from 'react-error-boundary'
 import { QueryClient } from '@tanstack/query-core'
 import * as MutationCacheModule from '../../../query-core/src/mutationCache'
